+++ src/neural_network/synapse.cpp
@@ -11,9 +11,9 @@
 
 
 
-long long int temp = 0;
+int64_t temp = 0;
 
-long long int synapse::synapse_id_generator = 0;
+int64_t synapse::synapse_id_generator = 0;
 
 synapse::synapse(neuron *input, neuron *output, float w, float step_size) {
     references = 0;
